pyproject.toml
[project]
name = "neural-sieve"
version = "0.1.0"
description = "The High-Signal External Memory for AI Influence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "watchdog>=4.0",
    "httpx>=0.27",
    "fastapi>=0.109",
    "uvicorn>=0.27",
    "jinja2>=3.1",
    "openai>=1.0",
    "mcp>=1.0",
    "pyyaml>=6.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "python-frontmatter>=1.1",
    "python-multipart>=0.0.9",
]

[project.scripts]
sieve = "sieve.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sieve"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
